+++ backend/identity-service/src/main/resources/db/migration/V21__create_analytics.sql
@@ -1,12 +1,12 @@
 CREATE TABLE analytics (
     analytic_id         SERIAL  PRIMARY KEY,
-    api_id              INT     NOT NULL REFERENCES apis(api_id) ON DELETE CASCADE,
+    api_id              BIGINT     NOT NULL REFERENCES apis(api_id) ON DELETE CASCADE,
     metric_date         DATE    NOT NULL,
-    total_requests      INT     NOT NULL DEFAULT 0,
-    successful_requests INT     NOT NULL DEFAULT 0,
-    failed_requests     INT     NOT NULL DEFAULT 0,
-    avg_latency_ms      INT     NOT NULL DEFAULT 0,
-    quota_consumed      INT     NOT NULL DEFAULT 0,
+    total_requests      BIGINT     NOT NULL DEFAULT 0,
+    successful_requests BIGINT     NOT NULL DEFAULT 0,
+    failed_requests     BIGINT     NOT NULL DEFAULT 0,
+    avg_latency_ms      BIGINT     NOT NULL DEFAULT 0,
+    quota_consumed      BIGINT     NOT NULL DEFAULT 0,
 
     UNIQUE (api_id, metric_date)    -- one row per API per day, upserted by the analytics job
 );
