+++ backend/identity-service/src/main/resources/db/migration/V2__create_api_plans.sql
@@ -1,8 +1,8 @@
 CREATE TABLE api_plans (
-    plan_id       SERIAL          PRIMARY KEY,
+    plan_id       BIGSERIAL          PRIMARY KEY,
     plan_name     VARCHAR(100)    NOT NULL UNIQUE,
-    rate_limit    INT             NOT NULL,   -- requests per second
-    quota_limit   INT             NOT NULL,   -- requests per day
+    rate_limit    BIGINT             NOT NULL,   -- requests per second
+    quota_limit   BIGINT             NOT NULL,   -- requests per day
     billing_cycle VARCHAR(20)     NOT NULL DEFAULT 'monthly',
     price         NUMERIC(10,2)   NOT NULL DEFAULT 0.00,
     description   TEXT,
