+++ backend/identity-service/src/main/resources/db/migration/V3__create_throttle_policies.sql
@@ -1,9 +1,9 @@
 CREATE TABLE throttle_policies (
-    tp_id         SERIAL       PRIMARY KEY,
+    tp_id         BIGSERIAL       PRIMARY KEY,
     policy_name   VARCHAR(100) NOT NULL UNIQUE,
     tier_type     VARCHAR(30)  NOT NULL,   -- APPLICATION / SUBSCRIPTION / API / OPERATION
-    request_count INT          NOT NULL,
-    unit_time     INT          NOT NULL,   -- window in seconds e.g. 60 = per minute
+    request_count BIGINT          NOT NULL,
+    unit_time     BIGINT          NOT NULL,   -- window in seconds e.g. 60 = per minute
     stop_on_quota BOOLEAN      NOT NULL DEFAULT true,
     description   TEXT,
