+++ backend/identity-service/src/main/resources/db/migration/V5__create_organizations.sql
@@ -1,8 +1,8 @@
 CREATE TABLE organizations (
-    org_id     SERIAL       PRIMARY KEY,
+    org_id     BIGSERIAL       PRIMARY KEY,
     org_name   VARCHAR(150) NOT NULL UNIQUE,
     domain     VARCHAR(100),
-    plan_id    INT          REFERENCES api_plans(plan_id) ON DELETE SET NULL,
+    plan_id    BIGINT          REFERENCES api_plans(plan_id) ON DELETE SET NULL,
     status     VARCHAR(20)  NOT NULL DEFAULT 'active',
     created_at TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP,
