+++ backend/identity-service/src/main/resources/db/migration/V6__create_users.sql
@@ -1,10 +1,10 @@
 CREATE TABLE users (
-    user_id       SERIAL       PRIMARY KEY,
+    user_id       BIGSERIAL       PRIMARY KEY,
     email         VARCHAR(150) NOT NULL UNIQUE,
     name          VARCHAR(100) NOT NULL,
     password_hash TEXT,                        -- NULL when using Keycloak SSO
-    role_id       INT          NOT NULL REFERENCES roles(role_id) ON DELETE RESTRICT,
-    org_id        INT          REFERENCES organizations(org_id)   ON DELETE SET NULL,
+    role_id       BIGINT          NOT NULL REFERENCES roles(role_id) ON DELETE RESTRICT,
+    org_id        BIGINT          REFERENCES organizations(org_id)   ON DELETE SET NULL,
     status        VARCHAR(20)  NOT NULL DEFAULT 'active',
     mfa_enabled   BOOLEAN      NOT NULL DEFAULT false,
     last_login_at TIMESTAMP,
